ref(core): Restructure hub exports
To make it a bit easier to replace/refactor them in coming steps.
