+++ packages/core/src/scope.ts
@@ -35,12 +35,6 @@ import { applyScopeDataToEvent } from './utils/applyScopeDataToEvent';
 const DEFAULT_MAX_BREADCRUMBS = 100;
 
 /**
- * The global scope is kept in this module.
- * When accessing this via `getGlobalScope()` we'll make sure to set one if none is currently present.
- */
-let globalScope: ScopeInterface | undefined;
-
-/**
  * Holds additional event information. {@link Scope.applyToEvent} will be
  * called by the client before an event will be sent.
  */
@@ -694,27 +688,6 @@ export class Scope implements ScopeInterface {
       this._notifyingListeners = false;
     }
   }
-}
-
-/**
- * Get the global scope.
- * This scope is applied to _all_ events.
- */
-export function getGlobalScope(): ScopeInterface {
-  if (!globalScope) {
-    globalScope = new Scope();
-  }
-
-  return globalScope;
-}
-
-/**
- * This is mainly needed for tests.
- * DO NOT USE this, as this is an internal API and subject to change.
- * @hidden
- */
-export function setGlobalScope(scope: ScopeInterface | undefined): void {
-  globalScope = scope;
 }
 
 function generatePropagationContext(): PropagationContext {
